Remove unused react-native Animated mock
This mock appears to be unnecessary for our current test suite. It also
leveraged an internal module path, which is brittle. It would break if
React Native decided to alter its path.

If we need to reinstate this mock in the future, we should likely
explore the following approach instead. https://git.io/JtNmU